[Refactor:Developer] Move performance warning to vue (#11550)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
There are currently no components that are actively used that are made
with vue.

### What is the new behavior?
The performance warning when there are too many queries made in
development is now rendered with vue. Additionally, this is a refactor
of the vue integration with twig to support individual components rather
than only entire pages.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
As much design feedback is possible would be great in this step, a lot
of this change will be the foundation for supporting vue, the developer
experience/style really matters here.

---------

Co-authored-by: Williams Chen <82904790+williamschen23@users.noreply.github.com>