[Feature:InstructorUI] Add Rubric to Upload/Download Gradeable (#11138)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
The gradeable upload and download was added without the rubric settings
being saved or uploaded.

Closes #10937 

### What is the new behavior?
The rubric section has been added to the upload and download portions of
this, reusing the add rubric code, so there is not much duplication
between the download/upload just rubrics, and download the whole
gradeable.