[Dependency] Upgrade Clang source to version 15.0.7 (#11584)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Currently the version of llvm that Submitty uses requires an older
version of cmake than what is supplied by apt. The version of llvm was
from Jan. 2018 (7.1.0) and there were no breaking changes until a
version released in Feb. 2022 (13.0.1), so that is what was chosen here.
Sample_CMakeLists.txt also had to updated to move
`cmake_minimum_required ` from 2.8 to 3.5 so that the newer version of
cmake would accept it.

### Other information
- It is probably going to be a breaking change when cmake no longer
supports 3.13.4, many of the directories will have to be moved around to
support the new llvm directory structure.
- This PR will not pass the ansible install test since until #11585 is
merged ansible will use the version of llvm specified by main.