[Bugfix:System] Run ansible install on PR (#11585)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Currently ansible-install in CI always clones main rather than the
correct commit, leading to false positives and negatives if main and the
current branch are different.

### What is the new behavior?
Ansible will clone based on `{{ github.sha }}` to make sure that the
repository contents match what is being tested.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
This PR will not pass tests until both it and #11584 are merged

---------

Co-authored-by: Troy-bailan-King <97757033+Troy-bailan-King@users.noreply.github.com>