[Refactor:System] Remove unused user connection management (#11582)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
There is no change in behavior with this PR, but a private member
property was removed since it is no longer used. The original purpose
was for preventing rebroadcasting discussion forum events to the user
that caused them, but now that client-side websocket broadcast is
disabled it has no purpose.