[Testing:Forum] Reset Forum Upon Test Failure (#11490)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [x] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?

In main, when a forum test fails the threads the test creates remain,
which breaks subsequent tests. This means that they must be manually
deleted.

### What is the new behavior?

Now, before each test Cypress logs out, checks whether the 3 test
threads exist, and then keeps on deleting threads with those names until
none exist, and then continues with the test.

### Other information?

This doesn't completely fix the forum test, there seems to be multiple
other issues with the tests which cause it to intermittently fail.

---------

Co-authored-by: Williams Chen <82904790+williamschen23@users.noreply.github.com>