[Bugfix:Forum] Delete Forum Threads via WebSocket (#11580)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

For deleting posts based on the forum WebSocket updates, only posts
within the current active thread are removed.

### What is the new behavior?

If any user is connected to an active thread that's been deleted,
they're redirected back to the forum page. For all other scenarios, the
thread instance is deleted from the list of threads.


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

Fixes #11565

Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>