[Bugfix:CourseMaterials] Routing Issue and Icon Type (#11594)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?

There is a routing issue where it tries to reload to a "banner" page or
location when in the file BannerController.php, every route has
"community_events" instead.

Also there is an airplane icon being used which doesn't fit so I changed
it to a palette icon.


These are all the routes in bannercontroller.php for reference: 
1.     #[Route("/community_events")]
2.     #[Route("/community_event/upload", methods: ["POST"])]
3.     #[Route("/community_events/delete", methods: ["POST"])]



### What is the new behavior?

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->