[Feature:RainbowGrades] Set points to readonly (#11569)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Currently, the point values of a gradeable can be modified in Rainbow
Grades such that they do not match the point value of the actual
gradeable. These values are only intended to be changed in gradeable
configs.

![image](https://github.com/user-attachments/assets/24d13160-e2fa-46e0-b594-c0e559928540)


### What is the new behavior?
The point value is read only, and shows the division of manual versus
autograded points, so long as there are non-zero values in the bucket.

![image](https://github.com/user-attachments/assets/a72fca95-d2db-44bf-9bb3-73540dba5e02)


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->