[Bugfix:Submission] Prevent Grade Inquiry Text Across Pages (#11588)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

When a user types into the text area within the grade inquiry section,
the key is based on a component with the ID tied to the current
gradable, which doesn't exist. Thus, the key is the same for all pages.


![image](https://github.com/user-attachments/assets/d71733e9-8474-4541-b60b-802c1210110d)


### What is the new behavior?

All draft content keys related to grade inquiry text areas are tied to
the specific course and are gradable. This local storage instance is
also cleared now for successful submissions.


![image](https://github.com/user-attachments/assets/3faa4fe4-6f71-40f8-86e8-ffceb3090586)


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
Testing requires a gradable to have an open grade inquiry window, and
any unsaved changes should not propagate to other locations now.

Fixes #11230