[Testing:System] Add Prettier Formatter for CSS Files (#11596)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently on StyleLint version 15, 76 rules that conflict with the
Prettier code formatter including the indentation rule are deprecated,
and will be completely removed when we migrate to version 16.

### What is the new behavior?
Before migrating to StyleLint V16, this PR adds Prettier to the project,
enforcing stricter overall CSS rules and ensuring that we are still
checking for indentation once we make the migration.

### Other information?
This PR intentionally does not apply formatting changes to avoid
polluting the diff. A separate PR will handle those updates after merge.
<!-- How did you test -->
To test, remove the ignored files found in site/.prettierignore, run 
npx prettier --write "**/*.css" in the Submitty/site directory, and see
the diffs.