[Feature:Developer] Test More Ansible Scripts (#11533)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Related to #10488, #10338, #10483, and #10489
Currently only the install process when using ansible is tested by
checking that the website correctly returns a response when called by
'curl'

### What is the new behavior?
Now there is a place for Cypress tests, and they can run after the
ansible install has completed. This will allow for easier testing of
ansible scripts, if/when more are added.