[Feature:CourseMaterials] Calculate how Rotate and Update Banner (#11595)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

This work is the start of the javascript logic for building out the
event advertisements PR. This sets up variables in localstorage:

1. open -> the banner is visible to the user or not
2. bannerArray -> the list of banners the user has yet to see
3. removedArray -> banners that the user will choose to remove
4. duckTalking -> this will be a different duck that signals if there
are new banners the user has yet to view

In this PR. We remove banners that a superuser may have removed from the
system so localstorage is update, and we add in new banners just
created.


PLEASE NOTE: There are no front end visible changes yet. This is just
the javascript preperation, so this would require just a code review of
any issues that may pop up down the road with these changes


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

---------

Co-authored-by: zacharymnp <56311867+zacharymnp@users.noreply.github.com>
Co-authored-by: Alex Lavallee <alex@lavallee.one>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>