[Bugfix:System] Null Section Registration (#11611)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
#Fixes 11471
Currently if a user drops a course, or is set to the null section, they
cannot re-register.

### What is the new behavior?
The self registration sections listed on the home screen now shows self
registration courses where the user is in the null section also, and the
user is able to re-register for the course.

---------

Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>