[Feature:System] Add IntroQuantumComputing to Vagrantfile (#11626)
### What is the current behavior?
Currently, Vagrantfile mounts the following optional companion
repositories to a shared directory in the VM: AnalysisTools,
AnalysisToolsTS, Lichen, RainbowGrades, Tutorial, CrashCourseCPPSyntax,
LichenTestData, DockerImages, DockerImagesRPI.

### What is the new behavior?
The file mounts an additional companion repository,
IntroQuantumComputing, if found in the same directory.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
I reloaded the VM to apply the changes in the Vagrantfile and changed a
folder name locally to ensure the changes were applied in the mirrored
directory.

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>