[Bugfix:TAGrading] Fix crashing due to large PDFs (#11447)
### Please check if the PR fulfils these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behaviour?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Fixes #11382

Currently, if a student uploads a large PDF file, attempting to view it
in the TA grading interface causes a fatal error. This results in the
grading interface crashing and an "Allowed memory size exhausted" error.

### What is the new behaviour?

- Implemented a file size limit for PDF annotation using
ini_get('memory_limit') / 5.
- If the file exceeds this limit, the system will prevent it from
loading in the PDF viewer and display a message prompting the user to
download it instead.
- Updated the frontend to handle the failure response and alert the user
accordingly.

<img width="1470" alt="Screenshot 2025-02-08 at 8 20 32 PM"
src="https://github.com/user-attachments/assets/410af512-13c2-46f1-a7c7-edf016777feb"
/>

### Other information?
<!-- Is this a breaking change? -->
This is not a breaking change.
<!-- How did you test -->
As a student, upload large PDFs. As an instructor, try to view these
PDFs in the TA grading interface. See the PDF, or the error message.

---------

Co-authored-by: Troy-bailan-King <97757033+Troy-bailan-King@users.noreply.github.com>
Co-authored-by: Justin Manion <jmanion32@gmail.com>