[Bugfix:Submission] Fix duplicate and incorrect auto refresh (#11627)
### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
#11623

AutoResults.twig and AutogradingResutsBox.twig share the same code. the
TA renders AutoResults.twig, and the user renders both. The overlap is
AutoResults.twig, so we can check the refresh only there.

In addition, when an instructor refreshes a page, check_refresh will
check if the instructors page refreshed. This is not intended as we want
to check the user in which they are regarding.

### What is the new behavior?
There is now only 1 refresh. We are also checking the user who got
refreshed instead of whoever hit refresh.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->