[Bugfix:InstructorUI] Fix peer grading panel (#11602)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
When an instructor visits
`/courses/<term>/<course>/gradeable/<gradeable>/update?nav_tab=3` and
changes which tabs should be visible to peer graders, the changes are
saved in the database, but upon retrieval they are not set properly,
leading to all Gradeable objects having everything set to show.

### What is the new behavior?
This PR both fetches the allowed tabs properly, and also ensures that
the tabs are not just removed from the header, but also not sent at all.


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->