[Bugfix:InstructorUI] Fix bulk uploads with different types (#11632)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
If a split by page bulk upload is started then a split by qr code is
started on the same gradeable the gradeable bulk upload page will be
stuck as a frog robot until the directories are manually deleted.

### What is the new behavior?
The uploads will work correctly showing both uploads in the preview.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->