[Bugfix:Submission] Fix Duplicate Popups for Large Files (#11635)
fixes #10579

### What is the current behavior?
When a student submits a file that exceeds a gradeable's limit, a
JavaScript alert is displayed. Then, when they submit a file with
correct size, a Submitty popup error with the same message as the JS
alert is displayed, alongside a Submitty success popup.

### What is the new behavior?
When a student submits a file that exceeds the gradeable's limit, a
Submitty error popup is displayed. Only the Submitty success popup is
displayed when a submission is successful.

### How to test
1) As a student, upload a file that exceeds a gradeable's limit and
click submit. See Submitty popup error.
2) Upload a file within the gradeable's limit and click submit. See
singular Submitty success popup.