[Refactor:TAGrading] Migrate TAGrading to typescript. (#11630)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
The following files use javascript currently.
* ta-grading.js
* ta-grading-keymap.js
* panel-selector-modal.js
* ta-grading-rubric.js
* ta-grading-rubric-conflict.js

### What is the new behavior?
The following files have been migrated mostly in-place to typescript.
* ta-grading.ts
* ta-grading-keymap.ts
* panel-selector-modal.ts
* ta-grading-rubric.ts
* ta-grading-rubric-conflict.ts

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
This change was made by renaming the original javascript files to
typescript and then going error-by-error to fix the missing types. The
most likely places that this PR will create regressions is with the
various http requests that have to be parsed, but all tests are passing
now and therefore more in-depth functionality testing is required.