[DevDependency] Bump phpunit to 10 (#11650)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
We are using phpunit 9.

### What is the new behavior?
Now upgraded to phpunit 10.

### Other information?
One of the transient dependencies, php-parser, was initially bumped to
version 5. The current version of PHPStan had an issue with it but if we
bump that too, then we should be able to increase php-parser to 5. So, I
forcefully downgraded back to version 4 in the lock file. If composer
keeps trying to auto-upgrade php-parser in the future, we may want to
pin it directly.

Unfortunately, setMethods was a deprecated method removed in 10 which
then addMethods added. However, addMethods was already deprecated in 10
too. This PR uses addMethods for now until we can find a better
replacement. The major plus side is that
https://github.com/Submitty/Submitty/pull/11642 and the associated
future PRs are no longer blocking this upgrade.

---------

Co-authored-by: williamschen23 <chenwill2005@gmail.com>
Co-authored-by: Alex Lavallee <73203142+lavalleeale@users.noreply.github.com>