[Bugfix:Autograding] Multiple docker debugging errors (#11697)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Fixes #11645
Currently, if there is a required docker image for an gradeable is
missing, autograding submissions notify both students and instructors in
the autograding results box. However, if multiple images are required
and are missing, the subsequent errors overwrite the details of the
previous error.

### What is the new behavior?
The subsequent errors are appended to a list of errors and looped
through to display each image error.

### Other information?
Current behavior

![before](https://github.com/user-attachments/assets/89ffa3b9-8b5f-4e42-862a-d3859e8d21a8)

New behavior

![after](https://github.com/user-attachments/assets/4a96ea15-6bd1-4a67-a6c7-f72011a82397)

---------

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>