[Bugfix:Developer] Fix flaky e2e test setup (#11709)
### What is the current behavior?
For the past few weeks, our complete set of cypress tests have been
failing on occasion for some PRs, even on re-run.
The hypothesis is that it is a timezone bug that relies on the assigned
data center/machine for github actions jobs (and perhaps on the time of
day when the job starts).

### What is the new behavior?
The timezone is now set explicitly for the postgress server for these
cypress jobs.
Hopefully this solves the bulk failing of cypress tests.