[Refactor:Forum] Move getEditPostContent to doctrine (#11239)
### What is the current behavior?
When getting the details needed to edit a post, we use vanilla PHP
instead of Doctrine.

### What is the new behavior?
When getting the details needed to edit a post, Doctrine will be used.
This allows the code and interactions with the database to be more
intuitive, and easier to write.

### Other information?
This will not affect functionality.

---------

Co-authored-by: DarthNyan <tatewhiteberg@gmail.com>
Co-authored-by: Tate Whiteberg <89590361+DarthNyan@users.noreply.github.com>
Co-authored-by: Garrett Gee <garrett.m.gee@gmail.com>
Co-authored-by: Justin Manion <jmanion32@gmail.com>