[UI/UX:Forum] Resizable and Scrollable Text Areas (#11704)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

Text areas on the discussion forum are automatically resized to fit the
entire content within the current screen, but this can be overwhelming
for large inputs.


![image](https://github.com/user-attachments/assets/2332744a-4a9b-4446-9670-78edf43c6ac2)

The markdown preview content is also limited to a max height of 300px,
which can be underwhelming for large inputs.


![image](https://github.com/user-attachments/assets/4a03ab2a-aac4-4544-9a5f-fdfac2d492b4)


### What is the new behavior?

Text areas on the discussion forum are still automatically resized;
however, additional functionality for resizing the vertical height and
scrolling accordingly has been added. When toggling between write and
preview mode, write-mode height and scroll positions are restored.


![image](https://github.com/user-attachments/assets/6bb447db-0bbf-40f8-9643-3c9ecdbc1f57)

Additionally, the markdown preview now has no height limit, displaying
the entirety of its content for readability measures.


![image](https://github.com/user-attachments/assets/c2ab2f1a-a79b-43ae-bc68-18acc53c5cb2)


### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

Fixes #10623

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>