[UI/UX:InstructorUI] Improved Gradeable Statistics UI/UX (#11713)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

The histogram related to gradable statistics lacks clarity in terms of
bucket size inputs and responsiveness on smaller devices on both the
histogram itself and the tab bar. Additionally, spacing across the
histogram and scatterplot inputs could be further improved for
readability, specifically on dark mode.


![image](https://github.com/user-attachments/assets/9d20a62b-4808-41a8-9276-c678e88448a5)


![image](https://github.com/user-attachments/assets/ba73360e-7c7f-4e4a-a0f2-90cc855b3a71)


![image](https://github.com/user-attachments/assets/a5baa6e4-7e4f-489e-99c2-c1af140f6ebd)



### What is the new behavior?

The bucket inputs were updated to display the placeholder `Enter bucket
size` with a button next to it with the text `Submit` for clarity
measures. Additionally, both the histogram and tab bar items have been
made responsive to enhance the user experience. Spacing for dark mode
users across the inputs, graphs, and statistics has been improved for
readability purposes.


![image](https://github.com/user-attachments/assets/d4d0c798-82cc-4f77-9ab5-b7c04fd1b0ce)


![image](https://github.com/user-attachments/assets/7369004a-ab6a-40c8-b1a9-705ff6b4122b)


![image](https://github.com/user-attachments/assets/c931f3b1-1c0c-482d-9ae1-c5754f55346b)



### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

Gradeable histograms can be viewed for assignments with grades available
after clicking the `Statistics & Charts` button within the details page
— fixes #10943.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>