[Bugfix:Autograding] Extend docker error to TAGrading (#11717)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently, having a missing docker image error only displays this error
for students on the homework submission side and does not display on the
page for TAGrading.

### What is the new behavior?
I reverted the changes to display the error in
AutogradingResultsBox.twig and only added it to AutoResults.twig. The
docker error now successfully shows for both TA grading and for student
submissions.

### Other information?
Before:

![image](https://github.com/user-attachments/assets/98242ab0-87d9-414d-8f7f-efac3c5feb19)

![image](https://github.com/user-attachments/assets/1e126527-d478-44f4-9f91-c7643f4b91be)

After:

![image](https://github.com/user-attachments/assets/f1e22295-eee7-49c6-b151-c7e5b4f7cebb)

![image](https://github.com/user-attachments/assets/0f37595e-b388-4a69-8e48-a66584cbd628)

<!-- How did you test -->
To test, I used the Docker Choice of Language gradeable in the
development course. I removed necessary docker images and submitted to
the gradeable.

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>