[Feature:Developer] Default Dates for sample data (#11701)
### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
The sample data has multiple dates, most of which are required. This
could greatly increase the number of lines that any couese.yml has. IN
addition, some of the fields in courses.yml were incorrect.

### What is the new behavior?
Changes some of the fields in courses.yml to be correct. Sets the
default date for every gradeable to be in the "OPENED" state with no
possible overdue submissions.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Barb Cutler <Barb Cutler>