[Bugfix:System] Remove docker images by id (#11357)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

The docker images that should exist on a worker machine are named in
several json files on the system. When the system is updated or
installed, it will automatically remove any images that aren't
specified. Currently, it does so by comparing the names of images on the
system with the names of images specified. This has a pitfall where if
an image has an alias on the system that isn't specified (two different
names pointing to the same image id), it will delete the image, and then
redownload it when it sees the required image isn't installed. While
this does result in every specified image being up-to-date on the system
at the end of installation, it results in a lot of unnecessary
redownloading. Since this is run in CI, the CI will need to download
10+GB from Docker's servers each run, which can trip the rate limit and
cause the run to fail in addition to the wasted bandwidth and time.

### What is the new behavior?

Determining the images to remove should be done by id instead of by
name. This will ensure that various aliases of the same image do not get
removed and the image will not need to be redownloaded.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

Fixes #11333 

#### Testing
You can add Docker images through DockerUI. The UI will display the
image once it has been installed. Add two images with two tags pointing
to the same id (submitty/gcc:latest, submitty/gcc:11). If you click
remove, you should be able to see that the primary image tag is removed
but the other tag remains. Logs are found at
/var/local/submitty/logs/docker/DATE.txt.

---------

Signed-off-by: Christopher Poon <christopherpoonc@gmail.com>
Co-authored-by: Christopher Poon <christopherpoonc@gmail.com>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>