[Bugfix:System] Handle Undefined Course in Email Subject (#11735)
### Why is this Change Important & Necessary?
fixes #11731
When Submitty sends an email, it assumes a course is defined and
associated with the sender. If no course is defined, an extra space
appears in the email subject line where the course label would normally
go:

```
reply-to: do-not-reply@vagrant
Subject: [Submitty ]: Submitty Email Verification
```

### What is the New Behavior?
Email header logic now checks if the course is defined to prevent the
unnecessary space:

```
reply-to: do-not-reply@vagrant
Subject: [Submitty]: Submitty Email Verification
```

### Other information
This is not a breaking change.