[UI/UX:Autograding] Improve Config Upload UX (#11139)
Fixes #10847 

### What is the current behavior?
When choosing an autograder config file to upload for a gradeable, the
user is not alerted if they have not actually uploaded the file. This
can cause confusion for users who go to other pages and thinking they
uploaded their config, when they didn't. Additionally, the UI is not
consistent with the rest of the site.
![old autograding configuration
page](https://github.com/user-attachments/assets/de96a289-a74f-44bf-9e50-1f11d07371c1)

### What is the new behavior?
The user is now alerted if they have an un-uploaded file, and given the
the chance to either upload it or cancel the upload.

This PR also enhances the config submission page by:
1) expanding the chosen file display length so it no longer gets cut off
2) changing the submit and cancel button appearances to be consistent
with the rest of the site
3) disabling the submit and cancel buttons when no files are uploaded so
the user can't click them and get an unnecessary error
4) moving the "Back to Submissions/Autograding" button to the top right
to follow the layout of other pages on Submitty.
![New Upload Autograding Config
page](https://github.com/user-attachments/assets/1673d15a-589e-4726-a845-13fcbc271ff0)


### Other information?
How to test:
1. Go to any gradeable and go into it's "Edit Gradeable Configuration"
via the pencil icon
2. Go to the Submissions/Autograding tab
3. Click "Upload a custom autograding configuration"
4. Upload a config file, and try to leave/refresh/close the page without
uploading or cancelling the upload.

This is not a breaking change.

---------

Co-authored-by: Justin Manion <jmanion32@gmail.com>
Co-authored-by: Williams Chen <82904790+williamschen23@users.noreply.github.com>